structured_outputs/pyproject.toml
[project]
name = "structured-outputs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "google-genai>=1.66.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-toon>=0.1.3",
]
